Replace `DateTime` with `CalDateTime` in `RecurrencePatternEvaluator` and related code. Key changes:

- Modifying the Until property of `RecurrencePattern` to use `CalDateTime`.
- Updating `ExceptionDates` to directly add `CalDateTime` instances.
- Refactoring methods like `GetIso8601WeekOfYear` to accept `CalDateTime` parameters.
- Adjusting serialization and deserialization processes for `CalDateTime`.
- Adjusting the `RecurrencePatternEvaluator` to work with `CalDateTime`.