@axunonb 2025-03-02
Here are the remaining members of `Ical.Net` that have a `DateTime` parameter or return type.
TBD: Any of them that should be updated, too?

- Has 'DateTime' parameter: AddTimeZone (Ical.Net.Calendar)
- Has 'DateTime' parameter: AddTimeZone (Ical.Net.Calendar)
- Has 'DateTime' parameter: AddLocalTimeZone (Ical.Net.Calendar)
- Returns 'DateTime': System.IConvertible.ToDateTime (Ical.Net.FreeBusyStatus)
- Returns 'DateTime': System.IConvertible.ToDateTime (Ical.Net.FrequencyType)
- Returns 'DateTime': System.IConvertible.ToDateTime (Ical.Net.FrequencyOccurrence)
- Has 'DateTime' parameter: AsCalDateTime (Ical.Net.Utility.DateUtil)
- Has 'DateTime' parameter: ToZonedDateTimeLeniently (Ical.Net.Utility.DateUtil)
- Has 'DateTime' parameter: FromTimeZoneToTimeZone (Ical.Net.Utility.DateUtil)
- Has 'DateTime' parameter: FromTimeZoneToTimeZone (Ical.Net.Utility.DateUtil)
- Returns 'DateTime': Truncate (Ical.Net.Utility.DateUtil)
- Has 'DateTime' parameter: Truncate (Ical.Net.Utility.DateUtil)
- Has 'DateTime' parameter: WeekOfMonth (Ical.Net.Utility.DateUtil)
- Returns 'DateTime': get_AsUtc (Ical.Net.DataTypes.CalDateTime)
- Returns 'DateTime': get_Value (Ical.Net.DataTypes.CalDateTime)
- Has 'DateTime' parameter: TruncateTimeToSeconds (Ical.Net.DataTypes.CalDateTime)
- Returns 'DateTime': System.IConvertible.ToDateTime (Ical.Net.DataTypes.PeriodKind)
- Returns 'DateTime': ToUtc (Ical.Net.DataTypes.UtcOffset)
- Has 'DateTime' parameter: ToUtc (Ical.Net.DataTypes.UtcOffset)
- Returns 'DateTime': ToLocal (Ical.Net.DataTypes.UtcOffset)
- Has 'DateTime' parameter: ToLocal (Ical.Net.DataTypes.UtcOffset)
- Has 'DateTime' parameter: FromLocalTimeZone (Ical.Net.CalendarComponents.VTimeZone)
- Has 'DateTime' parameter: FromSystemTimeZone (Ical.Net.CalendarComponents.VTimeZone)
- Has 'DateTime' parameter: FromDateTimeZone (Ical.Net.CalendarComponents.VTimeZone)