@minichma 2025-03-04
@axunonb 

> TBD: Any of them that should be updated, too? (Expecially considering GetOccurrences)

I'd say ideally most of them, especially the `GetOccurrences`, except maybe some internal ones and those intended for conversion to/from DateTime.