@axunonb 2025-03-04
Gonna create another PR for `GetOccurrences` updates.