@axunonb 2025-03-04
> Maybe check the performance impact of changing the internals of this class to CalDateTime. Could be measurable.

Yes, true, quite significant... You think the price is too high?

#### Current Main

```
// * Summary *

BenchmarkDotNet v0.14.0, Windows 11 (10.0.26100.3194)
13th Gen Intel Core i7-13700K, 1 CPU, 24 logical and 16 physical cores
.NET SDK 9.0.200
  [Host] : .NET 8.0.13 (8.0.1325.6609), X64 RyuJIT AVX2

Toolchain=InProcessNoEmitToolchain

| Method            | Mean       | Error     | StdDev    | Gen0     | Gen1     | Allocated |
|------------------ |-----------:|----------:|----------:|---------:|---------:|----------:|
| SerializeCalendar |   6.318 us | 0.0514 us | 0.0456 us |   1.4038 |   0.0229 |  21.61 KB |
| GetOccurrences    | 660.939 us | 2.7257 us | 2.5496 us | 143.5547 | 113.2813 |   2210 KB |

// * Hints *
Outliers
  RecurrenceBenchmark.SerializeCalendar: Toolchain=InProcessNoEmitToolchain -> 1 outlier  was  removed (6.58 us)

// * Legends *
  Mean      : Arithmetic mean of all measurements
  Error     : Half of 99.9% confidence interval
  StdDev    : Standard deviation of all measurements
  Gen0      : GC Generation 0 collects per 1000 operations
  Gen1      : GC Generation 1 collects per 1000 operations
  Allocated : Allocated memory per single operation (managed only, inclusive, 1KB = 1024B)
  1 us      : 1 Microsecond (0.000001 sec)

// * Diagnostic Output - MemoryDiagnoser *


// ***** BenchmarkRunner: End *****
Run time: 00:00:39 (39.13 sec), executed benchmarks: 2

Global total time: 00:00:39 (39.52 sec), executed benchmarks: 2
```

#### PR #742
```
// * Summary *

BenchmarkDotNet v0.14.0, Windows 11 (10.0.26100.3194)
13th Gen Intel Core i7-13700K, 1 CPU, 24 logical and 16 physical cores
.NET SDK 9.0.200
  [Host] : .NET 8.0.13 (8.0.1325.6609), X64 RyuJIT AVX2

Toolchain=InProcessNoEmitToolchain

| Method            | Mean       | Error     | StdDev    | Gen0     | Gen1    | Allocated  |
|------------------ |-----------:|----------:|----------:|---------:|--------:|-----------:|
| SerializeCalendar |   6.261 us | 0.0405 us | 0.0379 us |   1.4038 |  0.0229 |   21.61 KB |
| GetOccurrences    | 829.253 us | 2.2819 us | 1.7816 us | 156.2500 | 94.7266 | 2397.47 KB |

// * Hints *
Outliers
  RecurrenceBenchmark.GetOccurrences: Toolchain=InProcessNoEmitToolchain -> 3 outliers were removed, 5 outliers were detected (825.28 us, 826.41 us, 833.97 us..835.15 us)

// * Legends *
  Mean      : Arithmetic mean of all measurements
  Error     : Half of 99.9% confidence interval
  StdDev    : Standard deviation of all measurements
  Gen0      : GC Generation 0 collects per 1000 operations
  Gen1      : GC Generation 1 collects per 1000 operations
  Allocated : Allocated memory per single operation (managed only, inclusive, 1KB = 1024B)
  1 us      : 1 Microsecond (0.000001 sec)

// * Diagnostic Output - MemoryDiagnoser *


// ***** BenchmarkRunner: End *****
Run time: 00:00:43 (43.06 sec), executed benchmarks: 2

Global total time: 00:00:43 (43.44 sec), executed benchmarks: 2
```

#### Benchmark

```C#
[MemoryDiagnoser]
public class RecurrenceBenchmark
{
    private Calendar _calendar;

    [GlobalSetup]
    public void Setup()
    {
        _calendar = GenerateCalendarWithRecurrences();
    }

    private Calendar GenerateCalendarWithRecurrences()
    {
        var calendar = new Calendar();
        var dailyEvent = new CalendarEvent
        {
            Start = new CalDateTime(2025, 3, 1),
            End = null,
            RecurrenceRules = new List<RecurrencePattern>
            {
                new RecurrencePattern(FrequencyType.Daily, 1) { Count = 1000 }
            }
        };
        calendar.Events.Add(dailyEvent);
        return calendar;
    }

    [Benchmark]
    public void SerializeCalendar()
    {
        var serializer = new CalendarSerializer();
        var serializedCalendar = serializer.SerializeToString(_calendar);
    }

    [Benchmark]
    public void GetOccurrences()
    {
        var occurrences = _calendar.GetOccurrences().ToList();
    }
}
```