@minichma 2025-03-04
The problem with `CalDateTime` is that its a class while DateTime, DateOnly, etc. are structs. So for all the little operations in `RecurrenceEvaluator`, where tons of objects are allocated, this makes a difference. But the difference is lower than I'd have expected, so I wouldn't expect many complaints from users either way. You could either revert the change *just for the internals of `RecurrencePatternEvaluator`* (other classes won't be affected too much), or just leave it as is. Both is ok from my perspective. At some point we might either change `CalDateTime` to struct or move evaluation to different types as discussed in other threads. `ZonedDateTime` et al for instance are structs.