@axunonb 2025-03-04
@minichma I was actually hesitating decreasing performance.
Now I added the same benchmark test against v4.3.1, because I was confident that the new `IEnumerable` for occurrences had to be a lot better than the v4.3.1 `HashSet`. And it was, even when using `CalDateTime`, **kudos to your effort**.
So finally we should proceed (with a clear conscience) as you suggested.